+++ zebra-chain/src/sapling/arbitrary.rs
@@ -85,7 +85,9 @@ impl Arbitrary for Output {
                 cv: ExtendedPoint::generator().into(),
                 cm_u: sapling_crypto::note::ExtractedNoteCommitment::from_bytes(&[0u8; 32])
                     .unwrap(),
-                ephemeral_key: keys::EphemeralPublicKey(ExtendedPoint::generator().into()),
+                ephemeral_key: keys::EphemeralPublicKey(
+                    jubjub::AffinePoint::from(ExtendedPoint::generator()).to_bytes(),
+                ),
                 enc_ciphertext,
                 out_ciphertext,
                 zkproof,